chore(zero-cache): move options to separate file
